+++ run.sh
@@ -9,13 +9,19 @@ rm -f out.txt
 #python3 -m main 7 designs/test-designs/test_2.v --sv --milestone-file milestones/test_2.json --coi --strategy directed
 
 #or1200
-python3 -m main 60 or1200_subset.F --sv  --auto-plan --llm-provider deepseek --coi --strategy directed -t or1200_top
+#python3 -m main 60 or1200_subset.F --sv  --auto-plan --llm-provider deepseek --coi --strategy directed -t or1200_top
 #python3 -m main 30 or1200_subset.F --sv  --auto-plan --milestone-file milestones/or1200_subset.json --coi --strategy directed -t or1200_top
 #python3 -m main 30 or1200_subset.F --sv  --auto-plan --milestone-file milestones/or1200_p49.json --coi --strategy directed -t or1200_top
 
 #sub-test
 #python3 -m main 100 designs/test-designs/sub-test/sub.F --sv --auto-plan --llm-provider deepseek --coi --strategy directed -t top
 #python3 -m main 6 designs/test-designs/sub-test/sub.F --sv --milestone-file milestones/sub-test.json --coi --strategy directed
+
+#hack@dac18
+#python3 -m main 30 hackdac18.F --sv --auto-plan --llm-provider deepseek --coi --strategy directed -t top_wrapper
+#python3 -m main 30 hackdac18.F --sv  --auto-plan --milestone-file milestones/hackdac18/p1.json --coi --strategy directed -t top_wrapper
+python3 -m main 30 hackdac18.F --sv  --auto-plan --milestone-file milestones/hackdac18/p2.json --coi --strategy directed -t top_wrapper
+
 
 
 #python3 -m main 16 designs/test-designs/new_test/top_compat.sv --sv  -I designs/test-designs/new_test/include/ --auto-plan --llm-provider deepseek --coi --strategy directed